convert to float32 at output to normalize it.

+++ src/marqo/s2_inference/clip_utils.py
@@ -238,11 +238,10 @@ class CLIP:
 
 
     def _convert_output(self, output):
-
         if self.device == 'cpu':
-            return output.to(torch.float32).numpy()
+            return output.numpy()
         elif self.device.startswith('cuda'):
-            return output.cpu().to(torch.float32).numpy()
+            return output.cpu().numpy()
 
     @staticmethod
     def normalize(outputs):
@@ -414,7 +413,7 @@ class OPEN_CLIP(CLIP):
         self.image_input_processed = torch.stack([self.preprocess(_img).to(self.device) for _img in image_input])
 
         with torch.no_grad(), torch.autocast(device_type="cuda" if self.device.startswith("cuda") else "cpu"):
-            outputs = self.model.encode_image(self.image_input_processed)
+            outputs = self.model.encode_image(self.image_input_processed).to(torch.float32)
 
         if normalize:
             _shape_before = outputs.shape
@@ -431,7 +430,7 @@ class OPEN_CLIP(CLIP):
         text = self.tokenizer(sentence).to(self.device)
 
         with torch.no_grad(), torch.autocast(device_type="cuda" if self.device.startswith("cuda") else "cpu"):
-            outputs = self.model.encode_text(text)
+            outputs = self.model.encode_text(text).to(torch.float32)
 
         if normalize:
             _shape_before = outputs.shape
